Move slice iterators to existing modules
